Score表中新增字段status,answerCount,totalCount；创建批量添加用户答题记录、分数接口；解决mybatis与springboot版本兼容问题；将H2数据库存储模式改为文件模式。

+++ do/src/main/java/com/ScoopLink/userAnswers/dto/UserAnswer.java
@@ -43,7 +43,7 @@ public class UserAnswer {
     /**
      * 获得分数
      */
-    private Double scoreObtained;
+    private Integer scoreObtained;
 
     /**
      * 回答时间
